+++ CHANGES.md
@@ -2,3 +2,4 @@
 - Updated Company Collection to contain new User Password identification
 - Bugfix in Plugins on missing standard data
 - Show if an error occoured on Payment Links
+- Add list of all sent Payment Links